+++ knowYourCity/Classes/TidbitListViewController.m
@@ -79,7 +79,7 @@
      if (indexPath.row >= [self.tidbitList count]) {
      
      cell.textLabel.text = NSLocalizedString(@"Not Found.", @"Not found descriptive text.");
-     DLog(@"Couldn't convert section %d and row %d into a theme.", indexPath.section, indexPath.row);
+     DLog(@"Couldn't convert section %ld and row %ld into a theme.", (long)indexPath.section, (long)indexPath.row);
      
      } else {
      cell.textLabel.text = [tidbitDictionary objectForKey:@"title"];
@@ -139,7 +139,7 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
     
-    DLog(@"Tapped section %d and row %d", indexPath.section, indexPath.row);
+    DLog(@"Tapped section %ld and row %ld", (long)indexPath.section, (long)indexPath.row);
     
     // push a detail view, based on media type
